requirements.txt
numpy==1.14.3
opencv-python==3.4.0.12
Pillow==5.1.0
websocket-server==0.4
